Finished rest of controller classes

+++ src/main/java/com/mymedicalbills/restapi/repository/BillRepository.java
@@ -3,6 +3,6 @@ package com.mymedicalbills.restapi.repository;
 import com.mymedicalbills.restapi.entity.Bill;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BillRepository extends JpaRepository<Bill, Integer> {
+public interface BillRepository extends JpaRepository<Bill, Long> {
 
 }
